Add BaseAdmin extends SonataAdmin, refactor CRUD Admin, add filter on Node entity Admin
